savepoint: before implementing stateful transformer
